+++ CHANGELOG.rst
@@ -7,12 +7,14 @@ https://semver.org/spec/v2.0.0.html. The most important bits are:
 Changelog
 =========
 
-0.2.3 (2023-04-15)
+0.2.3 (2023-04-17)
 ------------------
 - fixed python 3.9 compatibility (traceback.format_exception syntax changed)
 - fixed deferred table copy when task is invalid (introduced with 0.2.2)
 - fixed mssql to not reflect full schema while renamings happen
 - fixed clearing of metadata tables for lazy tables and raw sql tables
+- fixed mssql synonym resolution when reading input table for pandas task
+- initial implementation of issue #62: make query canonical before hashing
 
 0.2.2 (2023-03-31)
 ------------------